move files according to class name
* drop unnecessary includes
