+++ eval/CMakeLists.txt
@@ -45,7 +45,6 @@ vespa_define_module(
     src/tests/instruction/dense_replace_type_function
     src/tests/instruction/dense_simple_expand_function
     src/tests/instruction/dense_simple_join_function
-    src/tests/instruction/dense_simple_map_function
     src/tests/instruction/dense_single_reduce_function
     src/tests/instruction/dense_tensor_create_function
     src/tests/instruction/dense_tensor_peek_function
@@ -62,6 +61,7 @@ vespa_define_module(
     src/tests/instruction/index_lookup_table
     src/tests/instruction/join_with_number
     src/tests/instruction/mixed_inner_product_function
+    src/tests/instruction/mixed_map_function
     src/tests/instruction/pow_as_map_optimizer
     src/tests/instruction/remove_trivial_dimension_optimizer
     src/tests/instruction/vector_from_doubles_function
