--- eval/src/tests/instruction/dense_simple_map_function/CMakeLists.txt
@@ -1,9 +1,0 @@
-# Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-vespa_add_executable(eval_dense_simple_map_function_test_app TEST
-    SOURCES
-    dense_simple_map_function_test.cpp
-    DEPENDS
-    vespaeval
-    GTest::GTest
-)
-vespa_add_test(NAME eval_dense_simple_map_function_test_app COMMAND eval_dense_simple_map_function_test_app)
